FINALY X2

+++ src/main/java/ru/hogwarts/school_HomeWork_29/service/FacultyService.java
@@ -2,6 +2,9 @@ package ru.hogwarts.school_HomeWork_29.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.hogwarts.school_HomeWork_29.Exception.FacultyNotFound;
+import ru.hogwarts.school_HomeWork_29.Exception.StudentNotFound;
+import ru.hogwarts.school_HomeWork_29.Exception.WrongNameOrColorException;
 import ru.hogwarts.school_HomeWork_29.Repository.FacultyRepository;
 import ru.hogwarts.school_HomeWork_29.model.Faculty;
 import java.util.*;
@@ -18,18 +21,30 @@ public class FacultyService {
     }
 
     public Faculty addFaculty(Faculty faculty) {
+        if (faculty.getName() == null || faculty.getName().isBlank()) {
+            throw new WrongNameOrColorException("Wrong name/color");
+        }
+        if (faculty.getColor() == null || faculty.getColor().isBlank()) {
+            throw new WrongNameOrColorException("Wrong name/color");
+        }
         return facultyRepository.save(faculty);
     }
 
     public Faculty findFaculty (long id) {
-        return facultyRepository.findById(id).orElseThrow();
+        return facultyRepository.findById(id).orElseThrow(() -> new FacultyNotFound("Faculty not found"));
     }
 
     public Faculty editFaculty (Faculty faculty) {
+        if (!facultyRepository.existsById(faculty.getId())) {
+            throw new FacultyNotFound("Faculty not found");
+        }
         return facultyRepository.save(faculty);
     }
 
     public void deleteFaculty (long id) {
+        if (!facultyRepository.existsById(id)) {
+            throw new FacultyNotFound("Faculty not found");
+        }
         facultyRepository.deleteById(id);
     }
